Refactored bundler utils
