+++ bundler.js
@@ -1,0 +1,1 @@
+export { extendConfig } from './src/utils/bundler.js';
