+++ main.js
@@ -1,4 +1,4 @@
 export * from './src/utils/common';
 export * from './src/utils/dom';
 export * from './src/render';
-export * from './src/utils/config';
+export * from './src/utils/bundler';
